changes to the simple fpm
* simplefpm uses nsCOM2 protocol for IPC
